Rename id-assign file
